JAMES-4228 Extract RecipientValidator into data-api
